requirements.txt
fastapi
uvicorn
litellm
python-dotenv
httpx
rich
pandas
numpy
scikit-learn
judgy